changing back
There's a conflict between animate.css/wow.js and my margins on mobile. The site width is only 100% when animations and wow.js are off. When I put "overflow-x: hidden;" in the html/body, all wow elements (except header, and all elements on chrome desktop) are hidden. 

I have to either disable all animations on mobile, or keep my opening animation with the white space on the right margin (on scroll).

Note: I got rid of the unneeded navbar css.